Rename scrollToc to toggleTocSubsections
The scrolling behavior was removed in 474fd24e81ddb39f60b09ca2ed69e67c3d4f8774
so this commit changes the name of scrollToc to reflect that.
